+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - PARTNER : Fix can't change the address if partner is an employee.
 - DEBT RECOVERY : Fix unable to run debt recovery accounting batch.
 - EXPENSE : complete my expense (Error message when no expense is selected)
+- Purchase Order Line: fix delivery state update on splitted stock move lines.
 
 ## [5.1.5] - 2019-03-30
 ## Features